backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.9.0",
    "diffusers>=0.34.0",
    "dspy>=2.6.27",
    "fastapi>=0.116.1",
    "opencv-python>=4.12.0.88",
    "requests>=2.32.4",
    "transformers>=4.53.3",
    "uvicorn>=0.35.0",
    "firebase-admin>=6.4.0",
    "pyyaml>=6.0",
    "aiohttp>=3.9.0",
    "python-multipart>=0.0.20",
  
]

[project.scripts]
generate-types = "scripts.generate_types:main"
